backend/requirements.txt
firebase-admin>=6.0.0
google-generativeai>=0.3.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
SpeechRecognition>=3.10.0
gTTS>=2.4.0
pydub>=0.25.1
edge-tts>=6.1.0 